updated service methods

+++ CarSpeedDataApp.Core/Services/ICarSpeedDataService.cs
@@ -6,7 +6,7 @@ public interface ICarSpeedDataService
 {
 	public IEnumerable<CarSpeedData> GetRange(DateTime? dateFrom, DateTime? dateTo, int? speed);
 
-	public IEnumerable<int> GetDay(DateTime date);
+	public IEnumerable<double> GetDay(DateTime date);
 
 	public void CreateList (List<CarSpeedData> dataList);
 }
